chore(#1841): archive 9 deprecated tool files + 14 test files (2512 LOC dead code)
Archive 7 legacy heartbeat tools + 2 legacy sync tools to _archives/:
- Heartbeat: check-heartbeats, get-heartbeat-state, get-offline-machines,
  get-warning-machines, register-heartbeat, start/stop-heartbeat-service
- Sync: sync-on-offline, sync-on-online

These files have 0 active imports from other source files. Superseded by:
- roosync_heartbeat (CONS-#443 Groupe 1, #1609 auto-heartbeat)
- roosync_sync_event (CONS-#443 Groupe 2)

Tests archived: 14 files (9 __tests__ + 5 unit tests)
Build: OK | Tests: 462 pass, 0 fail (9377 assertions)
