+++ servers/roo-state-manager/_archives/deprecated-tools-2026-05-04/heartbeat-service-wrappers.test.ts
