+++ servers/roo-state-manager/_archives/deprecated-tools-2026-05-04/sync-deprecated.test.ts
