+++ servers/roo-state-manager/_archives/deprecated-tools-2026-05-04/sync-on-offline.ts
@@ -1,0 +1,174 @@
+/**
+ * ARCHIVED 2026-05-04
+ * Superseded-by: roosync_sync_event
+ * Reason: CONS-#443 Groupe 2
+ * This file is dead code — 0 active imports from other source files.
+ * Retained for reference only. Safe to delete after 30 days (2026-05-04 + 30).
+ */
+
+/**
+ * @deprecated Superseded by roosync_sync_event. Kept for backward compatibility in registry.ts.
+ * Outil MCP : roosync_sync_on_offline
+ *
+ * Synchronise automatiquement les baselines lors de la détection offline d'une machine.
+ *
+ * @module tools/roosync/sync-on-offline
+ * @version 3.0.0
+ */
+
+import { z } from 'zod';
+import { getRooSyncService } from '../../services/lazy-roosync.js';
+import { HeartbeatServiceError } from '../../services/roosync/HeartbeatService.js';
+
+/**
+ * Schema de validation pour roosync_sync_on_offline
+ */
+export const SyncOnOfflineArgsSchema = z.object({
+  machineId: z.string()
+    .describe('Identifiant de la machine offline'),
+  createBackup: z.boolean().optional()
+    .describe('Créer une sauvegarde avant synchronisation (défaut: true)'),
+  dryRun: z.boolean().optional()
+    .describe('Mode simulation sans modification réelle (défaut: false)')
+});
+
+export type SyncOnOfflineArgs = z.infer<typeof SyncOnOfflineArgsSchema>;
+
+/**
+ * Schema de retour pour roosync_sync_on_offline
+ */
+export const SyncOnOfflineResultSchema = z.object({
+  success: z.boolean()
+    .describe('Indique si la synchronisation a réussi'),
+  machineId: z.string()
+    .describe('Identifiant de la machine'),
+  syncedAt: z.string()
+    .describe('Timestamp de la synchronisation (ISO 8601)'),
+  backupCreated: z.boolean()
+    .describe('Indique si une sauvegarde a été créée'),
+  backupPath: z.string().optional()
+    .describe('Chemin de la sauvegarde si créée'),
+  changes: z.object({
+    filesSynced: z.number()
+      .describe('Nombre de fichiers synchronisés'),
+    conflictsResolved: z.number()
+      .describe('Nombre de conflits résolus'),
+    decisionsCreated: z.number()
+      .describe('Nombre de décisions créées')
+  })
+    .describe('Détails des changements'),
+  message: z.string()
+    .describe('Message de confirmation')
+});
+
+export type SyncOnOfflineResult = z.infer<typeof SyncOnOfflineResultSchema>;
+
+/**
+ * Outil roosync_sync_on_offline
+ *
+ * Synchronise automatiquement les baselines lors de la détection offline d'une machine.
+ * Cette opération est typiquement appelée par le service de heartbeat.
+ *
+ * @param args Arguments validés
+ * @returns Résultat de la synchronisation
+ * @throws {HeartbeatServiceError} En cas d'erreur
+ */
+export async function roosyncSyncOnOffline(args: SyncOnOfflineArgs): Promise<SyncOnOfflineResult> {
+  try {
+    const {
+      machineId,
+      createBackup = true,
+      dryRun = false
+    } = args;
+
+    const rooSyncService = await getRooSyncService();
+    const heartbeatService = rooSyncService.getHeartbeatService();
+
+    // Vérifier que la machine est bien offline
+    const heartbeatData = heartbeatService.getHeartbeatData(machineId);
+    if (!heartbeatData || heartbeatData.status !== 'offline') {
+      throw new HeartbeatServiceError(
+        `La machine ${machineId} n'est pas offline`,
+        'MACHINE_NOT_OFFLINE'
+      );
+    }
+
+    // En mode simulation, retourner un résultat factice
+    if (dryRun) {
+      return {
+        success: true,
+        machineId,
+        syncedAt: new Date().toISOString(),
+        backupCreated: false,
+        changes: {
+          filesSynced: 0,
+          conflictsResolved: 0,
+          decisionsCreated: 0
+        },
+        message: `Mode simulation: synchronisation offline pour ${machineId}`
+      };
+    }
+
+    // Créer une sauvegarde si demandé
+    let backupPath: string | undefined;
+    if (createBackup) {
+      // Note: La logique de sauvegarde réelle serait implémentée ici
+      // Pour l'instant, nous simulons le chemin
+      backupPath = `roo-config/backups/offline-sync-${machineId}-${Date.now()}.json`;
+    }
+
+    // Note: La logique de synchronisation réelle serait implémentée ici
+    // Pour l'instant, nous simulons le résultat
+    const syncedAt = new Date().toISOString();
+
+    return {
+      success: true,
+      machineId,
+      syncedAt,
+      backupCreated: createBackup,
+      backupPath,
+      changes: {
+        filesSynced: 0,
+        conflictsResolved: 0,
+        decisionsCreated: 0
+      },
+      message: `Synchronisation offline effectuée pour ${machineId}`
+    };
+  } catch (error) {
+    if (error instanceof HeartbeatServiceError) {
+      throw error;
+    }
+
+    throw new HeartbeatServiceError(
+      `Erreur lors de la synchronisation offline: ${(error as Error).message}`,
+      'SYNC_OFFLINE_FAILED'
+    );
+  }
+}
+
+/**
+ * Métadonnées de l'outil pour l'enregistrement MCP
+ */
+export const syncOnOfflineToolMetadata = {
+  name: 'roosync_sync_on_offline',
+  description: 'Synchronise automatiquement les baselines lors de la détection offline d\'une machine. Cette opération est typiquement appelée par le service de heartbeat.',
+  inputSchema: {
+    type: 'object' as const,
+    properties: {
+      machineId: {
+        type: 'string',
+        description: 'Identifiant de la machine offline'
+      },
+      createBackup: {
+        type: 'boolean',
+        description: 'Créer une sauvegarde avant synchronisation (défaut: true)'
+      },
+      dryRun: {
+        type: 'boolean',
+        description: 'Mode simulation sans modification réelle (défaut: false)'
+      }
+    },
+    required: ['machineId'],
+    additionalProperties: false
+  }
+};
