+++ servers/roo-state-manager/_archives/deprecated-tools-2026-05-04/sync-on-online.ts
@@ -1,0 +1,191 @@
+/**
+ * ARCHIVED 2026-05-04
+ * Superseded-by: roosync_sync_event
+ * Reason: CONS-#443 Groupe 2
+ * This file is dead code — 0 active imports from other source files.
+ * Retained for reference only. Safe to delete after 30 days (2026-05-04 + 30).
+ */
+
+/**
+ * @deprecated Superseded by roosync_sync_event. Kept for backward compatibility in registry.ts.
+ * Outil MCP : roosync_sync_on_online
+ *
+ * Synchronise automatiquement les baselines lors du retour online d'une machine.
+ *
+ * @module tools/roosync/sync-on-online
+ * @version 3.0.0
+ */
+
+import { z } from 'zod';
+import { getRooSyncService } from '../../services/lazy-roosync.js';
+import { HeartbeatServiceError } from '../../services/roosync/HeartbeatService.js';
+
+/**
+ * Schema de validation pour roosync_sync_on_online
+ */
+export const SyncOnOnlineArgsSchema = z.object({
+  machineId: z.string()
+    .describe('Identifiant de la machine redevenue online'),
+  createBackup: z.boolean().optional()
+    .describe('Créer une sauvegarde avant synchronisation (défaut: true)'),
+  dryRun: z.boolean().optional()
+    .describe('Mode simulation sans modification réelle (défaut: false)'),
+  syncFromBaseline: z.boolean().optional()
+    .describe('Synchroniser depuis la baseline (défaut: true)')
+});
+
+export type SyncOnOnlineArgs = z.infer<typeof SyncOnOnlineArgsSchema>;
+
+/**
+ * Schema de retour pour roosync_sync_on_online
+ */
+export const SyncOnOnlineResultSchema = z.object({
+  success: z.boolean()
+    .describe('Indique si la synchronisation a réussi'),
+  machineId: z.string()
+    .describe('Identifiant de la machine'),
+  syncedAt: z.string()
+    .describe('Timestamp de la synchronisation (ISO 8601)'),
+  backupCreated: z.boolean()
+    .describe('Indique si une sauvegarde a été créée'),
+  backupPath: z.string().optional()
+    .describe('Chemin de la sauvegarde si créée'),
+  changes: z.object({
+    filesSynced: z.number()
+      .describe('Nombre de fichiers synchronisés'),
+    conflictsResolved: z.number()
+      .describe('Nombre de conflits résolus'),
+    decisionsCreated: z.number()
+      .describe('Nombre de décisions créées'),
+    offlineDuration: z.number().optional()
+      .describe('Durée offline en millisecondes')
+  })
+    .describe('Détails des changements'),
+  message: z.string()
+    .describe('Message de confirmation')
+});
+
+export type SyncOnOnlineResult = z.infer<typeof SyncOnOnlineResultSchema>;
+
+/**
+ * Outil roosync_sync_on_online
+ *
+ * Synchronise automatiquement les baselines lors du retour online d'une machine.
+ * Cette opération est typiquement appelée par le service de heartbeat.
+ *
+ * @param args Arguments validés
+ * @returns Résultat de la synchronisation
+ * @throws {HeartbeatServiceError} En cas d'erreur
+ */
+export async function roosyncSyncOnOnline(args: SyncOnOnlineArgs): Promise<SyncOnOnlineResult> {
+  try {
+    const {
+      machineId,
+      createBackup = true,
+      dryRun = false,
+      syncFromBaseline = true
+    } = args;
+
+    const rooSyncService = await getRooSyncService();
+    const heartbeatService = rooSyncService.getHeartbeatService();
+
+    // Vérifier que la machine est bien online
+    const heartbeatData = heartbeatService.getHeartbeatData(machineId);
+    if (!heartbeatData || heartbeatData.status !== 'online') {
+      throw new HeartbeatServiceError(
+        `La machine ${machineId} n'est pas online`,
+        'MACHINE_NOT_ONLINE'
+      );
+    }
+
+    // Calculer la durée offline si disponible
+    let offlineDuration: number | undefined;
+    if (heartbeatData.offlineSince) {
+      offlineDuration = Date.now() - new Date(heartbeatData.offlineSince).getTime();
+    }
+
+    // En mode simulation, retourner un résultat factice
+    if (dryRun) {
+      return {
+        success: true,
+        machineId,
+        syncedAt: new Date().toISOString(),
+        backupCreated: false,
+        changes: {
+          filesSynced: 0,
+          conflictsResolved: 0,
+          decisionsCreated: 0,
+          offlineDuration
+        },
+        message: `Mode simulation: synchronisation online pour ${machineId}`
+      };
+    }
+
+    // Créer une sauvegarde si demandé
+    let backupPath: string | undefined;
+    if (createBackup) {
+      // Note: La logique de sauvegarde réelle serait implémentée ici
+      // Pour l'instant, nous simulons le chemin
+      backupPath = `roo-config/backups/online-sync-${machineId}-${Date.now()}.json`;
+    }
+
+    // Note: La logique de synchronisation réelle serait implémentée ici
+    // Pour l'instant, nous simulons le résultat
+    const syncedAt = new Date().toISOString();
+
+    return {
+      success: true,
+      machineId,
+      syncedAt,
+      backupCreated: createBackup,
+      backupPath,
+      changes: {
+        filesSynced: 0,
+        conflictsResolved: 0,
+        decisionsCreated: 0,
+        offlineDuration
+      },
+      message: `Synchronisation online effectuée pour ${machineId}`
+    };
+  } catch (error) {
+    if (error instanceof HeartbeatServiceError) {
+      throw error;
+    }
+
+    throw new HeartbeatServiceError(
+      `Erreur lors de la synchronisation online: ${(error as Error).message}`,
+      'SYNC_ONLINE_FAILED'
+    );
+  }
+}
+
+/**
+ * Métadonnées de l'outil pour l'enregistrement MCP
+ */
+export const syncOnOnlineToolMetadata = {
+  name: 'roosync_sync_on_online',
+  description: 'Synchronise automatiquement les baselines lors du retour online d\'une machine. Cette opération est typiquement appelée par le service de heartbeat.',
+  inputSchema: {
+    type: 'object' as const,
+    properties: {
+      machineId: {
+        type: 'string',
+        description: 'Identifiant de la machine redevenue online'
+      },
+      createBackup: {
+        type: 'boolean',
+        description: 'Créer une sauvegarde avant synchronisation (défaut: true)'
+      },
+      dryRun: {
+        type: 'boolean',
+        description: 'Mode simulation sans modification réelle (défaut: false)'
+      },
+      syncFromBaseline: {
+        type: 'boolean',
+        description: 'Synchroniser depuis la baseline (défaut: true)'
+      }
+    },
+    required: ['machineId'],
+    additionalProperties: false
+  }
+};
